requirements.txt
absl-py==1.2.0
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
black==20.8b1
bleach==3.2.1
blinker==1.4
blis==0.4.1
cachetools==4.2.2
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clingo==5.5.1
colorama==0.4.4
coloredlogs==14.0
commonmark==0.9.1
cx-Freeze==6.4.2
cycler==0.10.0
cymem==2.0.6
Cython==0.29.32
decorator==4.4.2
defusedxml==0.6.0
distro==1.7.0
docopt==0.6.2
entrypoints==0.3
faiss-cpu==1.7.2
feedparser==6.0.2
filelock==3.0.12
flatbuffers==2.0
gast==0.4.0
gdown==3.12.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
htbuilder==0.6.0
huggingface-hub==0.0.8
humanfriendly==10.0
idna==2.10
importlib-metadata==4.12.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.6.4
iteration-utilities==0.11.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libclang==14.0.6
linkpreview==0.2.0
lxml==4.6.3
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.7
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
neo4j==4.2.1
nest-asyncio==1.4.3
networkx==2.5.1
newsapi-python==0.2.6
nltk==3.6.2
notebook==6.3.0
numpy==1.21.6
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.7
pandas==1.2.4
pandocfilters==1.4.3
parso==0.7.1
pathspec==0.8.1
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.0.1
pipreqs==0.4.10
plac==1.1.3
plotly==4.14.3
preshed==3.0.6
prometheus-client==0.10.1
prompt-toolkit==3.0.8
protobuf==3.15.8
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.5
pydeck==0.6.2
pygaggle==0.0.3.1
pygame==2.0.0
Pygments==2.7.2
pyjnius==1.4.2
pylint==2.6.0
Pympler==1.0.1
pyparsing==2.4.7
PyPDF2==2.10.0
pyreadline==2.1
pyrsistent==0.17.3
pyserini==0.9.4.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==303
pywinpty==1.0.1
pyzmq==20.0.0
qtconsole==5.1.0
QtPy==1.9.0
rank-bm25==0.2.2
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
retrying==1.3.3
rich==12.5.1
rsa==4.9
runipy==0.1.5
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.6.3
semver==2.13.0
Send2Trash==1.5.0
sentencepiece==0.1.94
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
soupsieve==2.2.1
spacy==2.2.4
srsly==1.0.5
st-annotated-text==3.0.0
streamlit==1.12.0
streamlit-agraph==0.0.35
tabula-py==2.4.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0rc0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
thinc==7.4.0
threadpoolctl==2.1.0
tldextract==3.1.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.45.0
traitlets==5.0.5
transformers==4.6.1
tweepy==3.10.0
typed-ast==1.4.1
typing-extensions==4.3.0
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.3
validators==0.18.2
wasabi==0.10.1
watchdog==2.0.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yarg==0.1.9
zipp==3.4.0